+++ tests/cell.cpp
@@ -1,23 +1,10 @@
 #include "main.h"
 #include "egcpool.h"
 
-TEST_CASE("MultibyteWidth") {
+TEST_CASE("Cell") {
   if(!enforce_utf8()){
     return;
   }
-  CHECK(0 == mbswidth(""));       // zero bytes, zero columns
-  CHECK(-1 == mbswidth("\x7"));   // single byte, non-printable
-  CHECK(1 == mbswidth(" "));      // single byte, one column
-  CHECK(5 == mbswidth("abcde"));  // single byte, one column
-  CHECK(1 == mbswidth("µ"));      // two bytes, one column
-  // FIXME take this back up to CHECK as soon as we figure out why some
-  // architectures seem to see this as a single column...
-  WARN(2 == mbswidth("\xf0\x9f\xa6\xb2"));     // four bytes, two columns
-  CHECK(6 == mbswidth("平仮名")); // nine bytes, six columns
-  CHECK(1 == mbswidth("\ufdfd")); // three bytes, ? columns, wcwidth() returns 1
-}
-
-TEST_CASE("Cell") {
   // common initialization
   notcurses_options nopts{};
   nopts.suppress_banner = true;
@@ -35,7 +22,20 @@ TEST_CASE("Cell") {
     cell_release(n_, &c);
   }
 
-SUBCASE("SetItalic") {
+  SUBCASE("MultibyteWidth") {
+    CHECK(0 == notcurses_mbswidth(nc_, ""));       // zero bytes, zero columns
+    CHECK(-1 == notcurses_mbswidth(nc_, "\x7"));   // single byte, non-printable
+    CHECK(1 == notcurses_mbswidth(nc_, " "));      // single byte, one column
+    CHECK(5 == notcurses_mbswidth(nc_, "abcde"));  // single byte, one column
+    CHECK(1 == notcurses_mbswidth(nc_, "µ"));      // two bytes, one column
+    // FIXME take this back up to CHECK as soon as we figure out why some
+    // architectures seem to see this as a single column...
+    WARN(2 == notcurses_mbswidth(nc_, "\xf0\x9f\xa6\xb2"));     // four bytes, two columns
+    CHECK(6 == notcurses_mbswidth(nc_, "平仮名")); // nine bytes, six columns
+    CHECK(1 == notcurses_mbswidth(nc_, "\ufdfd")); // three bytes, ? columns, wcwidth() returns 1
+  }
+
+  SUBCASE("SetItalic") {
   cell c = CELL_TRIVIAL_INITIALIZER;
     int dimy, dimx;
     notcurses_term_dim_yx(nc_, &dimy, &dimx);